// src/common/enum/mchg/mchgStage.enum.ts
export const enum MchgStage {
  CHOOSING_QUESTION = 'CHOOSING_QUESTION',
  ANSWERING_MAIN_QUESTION = 'ANSWERING_MAIN_QUESTION', // Chướng ngại vật
  ANSWERING_SUB_QUESTION = 'ANSWERING_SUB_QUESTION', // Câu hỏi phụ
  SHOWING_SUB_QUESTION_ANSWER = 'SHOWING_SUB_QUESTION_ANSWER',
  UPDATE_RESULTS = 'UPDATE_RESULTS',
}

export enum MchgGameState {
  RUNNING,
  NOT_RUNNING,
  PAUSED,
}